requirements.txt
streamlit
psycopg2-binary
pandas
plotly
apache-airflow
